requirements.txt
aniso8601==7.0.0
anyio==3.6.1
arabic-reshaper==2.1.3
argcomplete==2.0.0
asgiref==3.4.1
astroid==2.4.2
attrs==21.2.0
autobahn==21.3.1
Automat==20.2.0
autopep8==1.5.7
backports.cached-property==1.0.1
beautifulsoup4==4.10.0
black==21.12b0
CacheControl==0.12.10
cachy==0.3.0
certifi==2022.6.15
cffi==1.14.5
channels==3.0.3
charset-normalizer==2.0.9
cleo==0.8.1
click==8.0.3
clikit==0.6.2
constantly==15.1.0
coreapi==2.3.3
coreschema==0.0.4
coverage==6.2
crashtest==0.3.1
crispy-bootstrap5==0.6
cryptography==3.4.7
daphne==3.0.2
defusedxml==0.7.1
diff-match-patch==20200713
distlib==0.3.4
dj-database-url==0.5.0
Django==3.1.7
django-admin-relation-links==0.2.5
django-admin-volt==0.0.3
django-appconf==1.0.5
django-autoslug==1.9.8
django-bootstrap4==21.1
django-bootstrap5==21.3
django-classy-tags==2.0.0
django-cms==3.9.0
django-cors-headers==3.6.0
django-countries==7.2.1
django-crispy-forms==1.13.0
django-cryptography==1.0
django-datatable-view==2.1.6
django-debug-toolbar==3.2.4
django-faicon==0.1.0
django-filter==21.1
django-formtools==2.3
django-groups-manager==1.2.0
django-guardian==2.4.0
django-import-export==2.8.0
django-inline-svg==0.1.1
django-js-asset==1.2.2
django-jsonfield==1.4.1
django-model-utils==4.2.0
django-modelclone==0.7.1
django-mptt==0.13.4
django-ninja==0.19.1
django-phonenumber-field==6.1.0
django-rest-swagger==2.2.0
django-sekizai==2.0.0
django-svg-image-form-field==1.0.1
django-tables2==2.4.0
django-taggit==1.5.1
django-treebeard==4.5.1
djangocms-admin-style==2.0.2
djangorestframework==3.12.4
djangorestframework-simplejwt==5.1.0
drf-yasg==1.21.3
et-xmlfile==1.1.0
factory-boy==3.2.1
Faker==13.3.4
fastapi==0.79.0
filelock==3.6.0
flake8==3.8.4
future==0.18.2
graphene==2.1.9
graphene-django==2.14.0
graphql-core==2.3.2
graphql-relay==2.0.1
gunicorn==20.1.0
h11==0.13.0
html5lib==1.1
hyperlink==21.0.0
idna==3.2
importlib-metadata==4.11.3
incremental==21.3.0
inflection==0.5.1
iniconfig==1.1.1
isort==5.6.4
itypes==1.2.0
Jinja2==3.1.2
keyring==23.5.0
lazy-object-proxy==1.4.3
lockfile==0.12.2
Markdown==3.3.4
MarkupPy==1.14
MarkupSafe==2.1.1
mccabe==0.6.1
msgpack==1.0.3
mypy-extensions==0.4.3
numpy==1.23.0
odfpy==1.4.1
openapi-codec==1.3.2
openpyxl==3.0.10
packaging==21.3
pandas==1.4.3
pastel==0.2.1
pathspec==0.9.0
pexpect==4.8.0
phonenumbers==8.12.46
Pillow==8.3.1
pipx==1.1.0
pkginfo==1.8.2
platformdirs==2.4.0
plotly==5.6.0
pluggy==1.0.0
poetry==1.1.13
poetry-core==1.0.8
promise==2.3
psycopg2-binary==2.9.3
ptyprocess==0.7.0
py==1.11.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.7.0
pycparser==2.20
pydantic==1.9.1
pyflakes==2.2.0
Pygments==2.11.2
PyJWT==2.3.0
pylev==1.4.0
pylint==2.6.0
PyMySQL==1.0.2
pyOpenSSL==20.0.1
pyparsing==3.0.7
PyPDF2==1.26.0
pytest==6.2.5
pytest-base-url==1.4.2
pytest-django==4.5.2
pytest-factoryboy==2.1.0
pytest-html==3.1.1
pytest-metadata==2.0.1
pytest-selenium==3.0.0
pytest-variables==1.9.0
python-bidi==0.4.2
python-dateutil==2.8.2
python-decouple==3.4
python-multipart==0.0.5
pytz==2021.1
pyuca==1.2
PyYAML==6.0
reportlab==3.6.2
requests==2.26.0
requests-toolbelt==0.9.1
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.6
Rx==1.6.1
selenium==3.141.0
service-identity==21.1.0
shellingham==1.4.0
simplejson==3.17.6
singledispatch==3.7.0
six==1.13.0
sniffio==1.2.0
soupsieve==2.3.1
sqlparse==0.4.1
starlette==0.19.1
strawberry-graphql==0.104.2
strawberry-graphql-django==0.2.5
stripe==2.63.0
swapper==1.3.0
tablib==3.2.1
tenacity==6.3.1
text-unidecode==1.3
toml==0.10.2
tomli==1.2.3
tomlkit==0.10.1
tqdm==4.64.0
Twisted==21.2.0
txaio==21.2.1
typing_extensions==4.0.1
Unidecode==1.3.2
Unipath==1.1
uritemplate==4.1.1
urllib3==1.26.7
userpath==1.8.0
uvicorn==0.18.2
virtualenv==20.14.0
webdriver-manager==3.5.4
webencodings==0.5.1
whatever==0.6
whitenoise==5.3.0
wrapt==1.12.1
xhtml2pdf==0.2.5
xlrd==2.0.1
xlwt==1.3.0
yapf==0.30.0
zipp==3.7.0
zope.interface==5.4.0
